Consolidate `DataDocument` related handling into `deprecated` module
