+++ src/prefect/agent.py
@@ -12,12 +12,12 @@ import pendulum
 
 from prefect.blocks.core import Block
 from prefect.client import OrionClient, get_client
+from prefect.deprecated.data_documents import DataDocument
 from prefect.engine import propose_state
 from prefect.exceptions import Abort, ObjectNotFound
 from prefect.infrastructure import Infrastructure, InfrastructureResult, Process
 from prefect.logging import get_logger
 from prefect.orion.schemas.core import BlockDocument, FlowRun, WorkQueue
-from prefect.orion.schemas.data import DataDocument
 from prefect.orion.schemas.states import Failed, Pending
 from prefect.settings import PREFECT_AGENT_PREFETCH_SECONDS
