+++ src/prefect/orion/schemas/__init__.py
@@ -1,9 +1,8 @@
-from . import states, schedules, data, core, sorting, filters, responses, actions
+from . import states, schedules, core, sorting, filters, responses, actions
 
 __all__ = [
     "actions",
     "core",
-    "data",
     "filters",
     "responses",
     "schedules",
