+++ src/prefect/testing/standard_test_suites/blocks.py
@@ -4,7 +4,6 @@ from typing import Type
 from urllib.request import urlopen
 
 import pytest
-from PIL import Image
 
 from prefect.blocks.core import Block
