+++ src/prefect/testing/standard_test_suites/task_runners.py
@@ -10,7 +10,7 @@ import pytest
 
 from prefect import flow, task
 from prefect.client.schemas import State, TaskRun
-from prefect.orion.schemas.data import DataDocument
+from prefect.deprecated.data_documents import DataDocument
 from prefect.orion.schemas.states import StateType
 from prefect.task_runners import BaseTaskRunner, TaskConcurrencyType
 from prefect.testing.utilities import exceptions_equal
