+++ tests/client/test_orion_client.py
@@ -16,9 +16,9 @@ import prefect.exceptions
 from prefect import flow, tags
 from prefect.client import OrionClient, get_client
 from prefect.client.schemas import OrchestrationResult, Pending, Running, Scheduled
+from prefect.deprecated.data_documents import DataDocument
 from prefect.orion import schemas
 from prefect.orion.api.server import ORION_API_VERSION, create_app
-from prefect.orion.schemas.data import DataDocument
 from prefect.orion.schemas.schedules import IntervalSchedule
 from prefect.orion.schemas.states import StateType
 from prefect.settings import (
