+++ tests/orion/schemas/test_states.py
@@ -4,7 +4,7 @@ import pendulum
 import pydantic
 import pytest
 
-from prefect.orion.schemas.data import DataDocument
+from prefect.deprecated.data_documents import DataDocument
 from prefect.orion.schemas.states import (
     AwaitingRetry,
     Completed,
