+++ tests/test_futures.py
@@ -6,9 +6,9 @@ import pytest
 
 from prefect.client import OrionClient
 from prefect.client.schemas import Completed
+from prefect.deprecated.data_documents import DataDocument
 from prefect.flows import flow
 from prefect.futures import PrefectFuture, resolve_futures_to_data
-from prefect.orion.schemas.data import DataDocument
 from prefect.tasks import task
 from prefect.testing.utilities import assert_does_not_warn
