+++ tests/test_states.py
@@ -1,8 +1,8 @@
 import pytest
 
 from prefect.client.schemas import Completed, Failed, Pending, Running, State
+from prefect.deprecated.data_documents import DataDocument
 from prefect.futures import PrefectFuture
-from prefect.orion.schemas.data import DataDocument
 from prefect.orion.schemas.states import StateType
 from prefect.states import (
     is_state,
