+++ tests/test_tasks.py
@@ -11,6 +11,7 @@ import pytest
 from prefect import flow, get_run_logger, tags
 from prefect.client.schemas import State
 from prefect.context import PrefectObjectRegistry, TaskRunContext, get_run_context
+from prefect.deprecated.data_documents import DataDocument
 from prefect.engine import get_state_for_result
 from prefect.exceptions import (
     MappingLengthMismatch,
@@ -20,7 +21,6 @@ from prefect.exceptions import (
 from prefect.futures import PrefectFuture
 from prefect.orion import models
 from prefect.orion.schemas.core import TaskRunResult
-from prefect.orion.schemas.data import DataDocument
 from prefect.orion.schemas.states import StateType
 from prefect.tasks import Task, task, task_input_hash
 from prefect.testing.utilities import exceptions_equal, flaky_on_windows
